requirements.txt
# --extra-index-url https://download.pytorch.org/whl/cu124
torch==2.6.0
torchvision==0.21.0
transformers==4.57.0
peft==0.14.0
accelerate==1.3.0
fastapi==0.115.6
tensorboard==2.18.0
uvicorn==0.34.0
json-numpy==2.1.1
safetensors==0.5.3
scipy==1.15.1
einops==0.8.1
timm==1.0.15
mmengine==0.10.5
pyarrow==19.0.0
h5py==3.12.1
mediapy==1.2.1
num2words==0.5.14
av==14.1.0
wandb==0.19.1
websockets==15.0
msgpack-numpy==0.4.8
# flash-attn==2.5.6
tensorflow==2.18.0
tensorflow-datasets==4.9.6
ninja==1.13.0

